wallet: split wallet store implementations to specific files
